front: update advertisement crud page

+++ backend/controller/advertisement.js
@@ -15,10 +15,41 @@ const {
   DELETE_ADVERTISEMENT,
 } = require("../config/message");
 
-//@description     get advertisements list
+//@description     get advertisements list for dashboard
 //@route           GET /api/advertisements
+//@access          protected
+router.get("/", authMiddleware, async (req, res) => {
+  const { user } = req;
+  try {
+    let { perPage, page, sort, search } = req.query;
+    let where =
+      user.id === "EMPLOYER"
+        ? { isShared: true, userId: user.id }
+        : { isShared: true };
+    console.log(where);
+    const result = await getPaginatedResults({
+      model: "Advertisement",
+      page: page ?? 1,
+      perPage: perPage ?? 10,
+      searchColumns: ["title", "companyName", "location"],
+      searchText: search,
+      sort,
+      where,
+      include: {
+        contract: true,
+        category: true,
+      },
+    });
+    res.send(result);
+  } catch (e) {
+    res.status(500).json({ message: [ERROR_500] });
+  }
+});
+
+//@description     get advertisements list
+//@route           GET /api/advertisements/list
 //@access          public
-router.get("/", async (req, res) => {
+router.get("/list", async (req, res) => {
   try {
     let { perPage, page, sort, search } = req.query;
     const result = await getPaginatedResults({
@@ -45,9 +76,6 @@ router.get("/", async (req, res) => {
 //@access          protected EMPLOYER
 router.post("/", authMiddleware, async (req, res) => {
   const { user, body } = req;
-  if (user.role !== "EMPLOYER")
-    return res.status(403).json({ message: ["access denied"] });
-
   const { error } = validateCreate(body);
   if (error) return res.status(400).send({ message: error });
 
@@ -64,7 +92,7 @@ router.post("/", authMiddleware, async (req, res) => {
 //@description     update advertisement
 //@route           Put /api/advertisements
 //@access          protected EMPLOYER
-router.put("/:id", [authMiddleware, isEmployer], async (req, res) => {
+router.put("/:id", authMiddleware, async (req, res) => {
   const { user, body, params } = req;
   if (!Number(params.id))
     return res.status(404).json({ message: ["no url params provided"] });
@@ -78,6 +106,11 @@ router.put("/:id", [authMiddleware, isEmployer], async (req, res) => {
     });
     if (!isExisting) return res.status(404).json({ message: [NOT_FOUND] });
 
+    if (isExisting.userId !== user.id)
+      return res.status(403).json({
+        message: "هر کاربر فقط آگهی های مربوط به خود را می تواند ویرایش کند",
+      });
+
     const result = await prisma.Advertisement.update({
       where: { id: parseInt(params.id) },
       data: body,
@@ -88,12 +121,22 @@ router.put("/:id", [authMiddleware, isEmployer], async (req, res) => {
   }
 });
 
-router.delete("/:id", [authMiddleware, isEmployer], async (req, res) => {
+router.delete("/:id", authMiddleware, async (req, res) => {
   const { params } = req;
   if (!Number(params.id))
     return res.status(404).json({ message: ["no url params provided"] });
 
   try {
+    const isExisting = await prisma.Advertisement.findUnique({
+      where: { id: parseInt(params.id) },
+    });
+    if (!isExisting) return res.status(404).json({ message: [NOT_FOUND] });
+
+    if (isExisting.userId !== user.id)
+      return res.status(403).json({
+        message: "هر کاربر فقط آگهی های مربوط به خود را می تواند حذف کند",
+      });
+
     await prisma.Advertisement.delete({ where: { id: parseInt(params.id) } });
     res.json({ message: [DELETE_ADVERTISEMENT] });
   } catch (e) {
